requirements.txt
absl_py==2.1.0
accelerate==0.34.2
datasets==2.21.0
deepspeed==0.16.2
gin_config==0.5.0
gluonts
matplotlib==3.10.0
numpy
pandas==2.2.3
PyYAML==6.0.2
reformer_pytorch==1.4.4
torch==2.4.1
tqdm==4.67.1
transformers
plotly==5.24.1
kaleido
nbformat
ipykernel
autogluon.timeseries
chronos-forecasting @ git+https://github.com/amazon-science/chronos-forecasting.git@main
chronos-forecasting[training] @ git+https://github.com/amazon-science/chronos-forecasting.git@main
